alert listo

+++ src/components/UserCard.jsx
@@ -1,17 +1,42 @@
 import './styles/userCard.css'
+import Swal from 'sweetalert2';
 
-const UserCard = ({ user, deleteUserById, setUpdateInfo,setFormClose }) => {
+const UserCard = ({ user, deleteUserById, setUpdateInfo, setFormClose }) => {
+
+
+
+
+
+
+
+
   const handleDelete = () => {
-    deleteUserById(user.id);
-  };
-
+    Swal.fire({
+      title: 'Are you sure to delete this user?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteUserById(user.id);
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
+  }
 
   const handleUpdate = () => {
     setUpdateInfo(user)
     setFormClose(false)
   }
 
- 
+
 
   return (
     <article className="user">
